Some mini fixes..
Ok, fixed the Key generation a bit more acceptable for strict compilers.
Also, implemented basic constructors for mesh3D...

+++ Mesh/btree.h
@@ -29,7 +29,7 @@ typedef struct _bintree_node {
 
     unsigned long depth;
 
-    unsigned long key;
+    unsigned int key;
     btree_data_t stuff;
 } bintree_node;
 typedef bintree_node* BTNode;
